requirements/tsting.txt
pytest>=2.6.4
pytest-cov>=1.8.1
pytest-django>=2.7.0
coveralls>=0.4.4
